Add bcrypt, slight refactor of authentication end-to-end
