+++ package.json
@@ -23,8 +23,10 @@
     "@hookform/resolvers": "^2.9.8",
     "@prisma/client": "^4.4.0",
     "@tanstack/react-query": "^4.2.3",
+    "@types/bcrypt": "^5.0.0",
     "@types/formidable": "^2.0.5",
     "babel-plugin-superjson-next": "^0.4.4",
+    "bcrypt": "^5.1.0",
     "formidable": "v3",
     "next": "12.2.0",
     "next-auth": "^4.14.0",
