+++ src/pages/api/submission.ts
@@ -7,7 +7,7 @@ import {
 import { buildResponse } from "@/common/lib/ResponseBuilder";
 import { withValidation } from "@/common/lib/ApiValidator";
 import { Submission, submissionSchema } from "@/challenges/schemas/submission";
-import { ResponseWithData } from "@/common/types/ResponseWithData";
+import { ResponseWithData } from "@/common/dto/ResponseWithData";
 import { getSession } from "next-auth/react";
 
 export default nc<NextApiRequest, NextApiResponse>({
